+++ OUDS/Core/OUDS/Tests/OUDSTheme/MockTheme/MockTheme+DimensionSemanticTokens.swift
@@ -20,5 +20,29 @@ extension MockTheme {
 
     static let mockThemeDimensionSemanticToken: DimensionSemanticToken = 007
 
-    override var dimensionDensityDefault: DimensionSemanticToken { Self.mockThemeDimensionSemanticToken }
+    override var dimension0: DimensionSemanticToken { Self.mockThemeDimensionSemanticToken }
+    override var dimension10: DimensionSemanticToken { Self.mockThemeDimensionSemanticToken }
+    override var dimension20: DimensionSemanticToken { Self.mockThemeDimensionSemanticToken }
+    override var dimension30: DimensionSemanticToken { Self.mockThemeDimensionSemanticToken }
+    override var dimension40: DimensionSemanticToken { Self.mockThemeDimensionSemanticToken }
+    override var dimension50: DimensionSemanticToken { Self.mockThemeDimensionSemanticToken }
+    override var dimension60: DimensionSemanticToken { Self.mockThemeDimensionSemanticToken }
+    override var dimension70: DimensionSemanticToken { Self.mockThemeDimensionSemanticToken }
+    override var dimension80: DimensionSemanticToken { Self.mockThemeDimensionSemanticToken }
+    override var dimension90: DimensionSemanticToken { Self.mockThemeDimensionSemanticToken }
+    override var dimension100: DimensionSemanticToken { Self.mockThemeDimensionSemanticToken }
+    override var dimension110: DimensionSemanticToken { Self.mockThemeDimensionSemanticToken }
+    override var dimension120: DimensionSemanticToken { Self.mockThemeDimensionSemanticToken }
+    override var dimension130: DimensionSemanticToken { Self.mockThemeDimensionSemanticToken }
+    override var dimension140: DimensionSemanticToken { Self.mockThemeDimensionSemanticToken }
+    override var dimension150: DimensionSemanticToken { Self.mockThemeDimensionSemanticToken }
+    override var dimension160: DimensionSemanticToken { Self.mockThemeDimensionSemanticToken }
+    override var dimension170: DimensionSemanticToken { Self.mockThemeDimensionSemanticToken }
+    override var dimension180: DimensionSemanticToken { Self.mockThemeDimensionSemanticToken }
+    override var dimension190: DimensionSemanticToken { Self.mockThemeDimensionSemanticToken }
+    override var dimension200: DimensionSemanticToken { Self.mockThemeDimensionSemanticToken }
+    override var dimension210: DimensionSemanticToken { Self.mockThemeDimensionSemanticToken }
+    override var dimension220: DimensionSemanticToken { Self.mockThemeDimensionSemanticToken }
+    override var dimension230: DimensionSemanticToken { Self.mockThemeDimensionSemanticToken }
+    override var dimension240: DimensionSemanticToken { Self.mockThemeDimensionSemanticToken }
 }